DO NOT REVERT: Changing to new library for robots.txt to allow me to not have to manage the nrobots myself. Lib not acting as expected https://code.google.com/p/robotstxt/issues/detail?id=6

+++ Abot/Crawler/PoliteWebCrawler.cs
@@ -41,11 +41,11 @@ namespace Abot.Crawler
 
         public override CrawlResult Crawl(Uri uri, CancellationTokenSource cancellationTokenSource)
         {
-            int robotsDotTextCrawlDelayInSecs = 0;
-            int robotsDotTextCrawlDelayInMillisecs = 0;
+            long robotsDotTextCrawlDelayInSecs = 0;
+            long robotsDotTextCrawlDelayInMillisecs = 0;
 
-            int maxRobotsDotTextCrawlDelayInSeconds = _crawlContext.CrawlConfiguration.MaxRobotsDotTextCrawlDelayInSeconds;
-            int maxRobotsDotTextCrawlDelayInMilliSecs = maxRobotsDotTextCrawlDelayInSeconds * 1000;
+            long maxRobotsDotTextCrawlDelayInSeconds = _crawlContext.CrawlConfiguration.MaxRobotsDotTextCrawlDelayInSeconds;
+            long maxRobotsDotTextCrawlDelayInMilliSecs = maxRobotsDotTextCrawlDelayInSeconds * 1000;
 
             //Load robots.txt
             if (_crawlContext.CrawlConfiguration.IsRespectRobotsDotTextEnabled)
