+++ Abot/Poco/CrawlConfiguration.cs
@@ -177,7 +177,7 @@ namespace Abot.Poco
         /// <summary>
         /// The number of milliseconds to wait in between http requests to the same domain.
         /// </summary>
-        public int MinCrawlDelayPerDomainMilliSeconds { get; set; }
+        public long MinCrawlDelayPerDomainMilliSeconds { get; set; }
         
         #endregion
     }
